refactor: move constants to folder
