+++ frontend/src/constants/stack.ts
