+++ frontend/tsconfig.json
@@ -32,7 +32,8 @@
       "@types/*": ["./src/types/*"],
       "@providers/*": ["./src/providers/*"],
       "@i18n/*": ["./src/i18n/*"],
-      "@pages/*": ["./src/pages/*"]
+      "@pages/*": ["./src/pages/*"],
+      "@constants/*": ["./src/constants/*"]
     }
   },
   "include": ["src"],
