+++ frontend/vite.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
       '@providers': path.resolve(__dirname, 'src/providers'),
       '@i18n': path.resolve(__dirname, 'src/i18n'),
       '@pages': path.resolve(__dirname, 'src/pages'),
+      '@constants': path.resolve(__dirname, 'src/constants'),
     },
   },
 });
